requirements.txt
Flask==3.1.0
PyJWT==2.10.1
bcrypt==4.2.1
